@@ -1,6 +1,6 @@
 {
     "name": "awg-app",
-    "version": "0.11.5",
+    "version": "0.11.6",
     "license": "MIT",
     "author": {
         "name": "Stefan Münnich",
@@ -82,16 +82,16 @@
         "deploy:ci": "ng deploy --no-build --message=\"Release $npm_package_name (v$npm_package_version) on gh-pages\""
     },
     "dependencies": {
-        "@angular/animations": "^17.1.3",
-        "@angular/common": "^17.1.3",
-        "@angular/compiler": "^17.1.3",
-        "@angular/core": "^17.1.3",
-        "@angular/forms": "^17.1.3",
-        "@angular/localize": "^17.1.3",
-        "@angular/platform-browser": "^17.1.3",
-        "@angular/platform-browser-dynamic": "^17.1.3",
-        "@angular/platform-server": "^17.1.3",
-        "@angular/router": "^17.1.3",
+        "@angular/animations": "^17.2.1",
+        "@angular/common": "^17.2.1",
+        "@angular/compiler": "^17.2.1",
+        "@angular/core": "^17.2.1",
+        "@angular/forms": "^17.2.1",
+        "@angular/localize": "^17.2.1",
+        "@angular/platform-browser": "^17.2.1",
+        "@angular/platform-browser-dynamic": "^17.2.1",
+        "@angular/platform-server": "^17.2.1",
+        "@angular/router": "^17.2.1",
         "@codemirror/legacy-modes": "^6.3.3",
         "@fortawesome/angular-fontawesome": "^0.14.1",
         "@fortawesome/fontawesome-svg-core": "^6.5.1",
@@ -118,14 +118,14 @@
         "zone.js": "~0.14.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^17.1.3",
+        "@angular-devkit/build-angular": "^17.2.0",
         "@angular-eslint/builder": "^17.2.1",
         "@angular-eslint/eslint-plugin": "^17.2.1",
         "@angular-eslint/eslint-plugin-template": "^17.2.1",
         "@angular-eslint/schematics": "^17.2.1",
         "@angular-eslint/template-parser": "^17.2.1",
-        "@angular/cli": "^17.1.3",
-        "@angular/compiler-cli": "^17.1.3",
+        "@angular/cli": "^17.2.0",
+        "@angular/compiler-cli": "^17.2.1",
         "@commitlint/cli": "^18.6.1",
         "@commitlint/config-angular": "^18.6.1",
         "@compodoc/compodoc": "^1.1.23",
