@@ -1,15 +1,15 @@
 // THIS IS AN AUTO-GENERATED FILE. DO NOT CHANGE IT MANUALLY!
-// Generated last time on Wed, Feb 14, 2024  1:15:38 PM
+// Generated last time on Mon, Feb 19, 2024  6:33:56 PM
 
 /**
  * The latest version of the AWG App
  */
-export const appVersion = '0.11.5';
+export const appVersion = '0.11.6';
 
 /**
  * The release date of the latest version of the AWG App
  */
-export const appVersionReleaseDate = '14. Februar 2024';
+export const appVersionReleaseDate = '19. Februar 2024';
 
 /**
  * The URL of the AWG App
