@@ -9,7 +9,7 @@ export const EDITION_FIRM_SIGNS_DATA = {
      * The path to the assets image of a firm sign: JE No. 12, 12 Lines.
      */
     FIRM_JE_NO_2_LIN_12: {
-        route: 'assets/img/edition/firm-signs/firm_je_no2_lin12.png',
+        route: 'assets/img/edition/firm-signs/je/no2_lin12/firm_je_no2_lin12.png',
         short: '[JE] | No. 2 | 12 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 2 | 12 linig.',
     },
@@ -18,7 +18,7 @@ export const EDITION_FIRM_SIGNS_DATA = {
      * The path to the assets image of a firm sign: JE No. 12, 12 Lines (variant "b").
      */
     FIRM_JE_NO_2_LIN_12_B: {
-        route: 'assets/img/edition/firm-signs/firm_je_no2_lin12_b.png',
+        route: 'assets/img/edition/firm-signs/je/no2_lin12/firm_je_no2_lin12_b.png',
         short: '[JE] | No. 2 | 12 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 2 | 12 linig.',
     },
@@ -27,7 +27,7 @@ export const EDITION_FIRM_SIGNS_DATA = {
      * The path to the assets image of a firm sign: JE No. 3, 14 Lines.
      */
     FIRM_JE_NO_3_LIN_14: {
-        route: 'assets/img/edition/firm-signs/firm_je_no3_lin14.png',
+        route: 'assets/img/edition/firm-signs/je/no3_lin14/firm_je_no3_lin14.png',
         short: '[JE] | No. 3 | 14 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 3 | 14 linig.',
     },
@@ -35,8 +35,8 @@ export const EDITION_FIRM_SIGNS_DATA = {
     /**
      * The path to the assets image of a firm sign: JE No. 5, 18 Lines.
      */
-    FIRM_JE_NO_5_LIN_18B: {
-        route: 'assets/img/edition/firm-signs/firm_je_no5_lin18b.png',
+    FIRM_JE_NO_5_LIN_18_B: {
+        route: 'assets/img/edition/firm-signs/je/no5_lin18/firm_je_no5_lin18_b.png',
         short: '[JE] | No. 5 | 18 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 5 | 18 linig.',
     },
@@ -44,8 +44,8 @@ export const EDITION_FIRM_SIGNS_DATA = {
     /**
      * The path to the assets image of a firm sign: JE No. 5, 18 Lines.
      */
-    FIRM_JE_NO_5_LIN_18C: {
-        route: 'assets/img/edition/firm-signs/firm_je_no5_lin18c.png',
+    FIRM_JE_NO_5_LIN_18_C: {
+        route: 'assets/img/edition/firm-signs/je/no5_lin18/firm_je_no5_lin18_c.png',
         short: '[JE] | No. 5 | 18 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 5 | 18 linig.',
     },
@@ -53,8 +53,8 @@ export const EDITION_FIRM_SIGNS_DATA = {
     /**
      * The path to the assets image of a firm sign: JE No. 5, 18 Lines.
      */
-    FIRM_JE_NO_5_LIN_18D: {
-        route: 'assets/img/edition/firm-signs/firm_je_no5_lin18d.png',
+    FIRM_JE_NO_5_LIN_18_D: {
+        route: 'assets/img/edition/firm-signs/je/no5_lin18/firm_je_no5_lin18_d.png',
         short: '[JE] | No. 5 | 18 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 5 | 18 linig.',
     },
@@ -63,7 +63,7 @@ export const EDITION_FIRM_SIGNS_DATA = {
      * The path to the assets image of a firm sign: JE No. 6, 20 Lines.
      */
     FIRM_JE_NO_6_LIN_20: {
-        route: 'assets/img/edition/firm-signs/firm_je_no6_lin20.png',
+        route: 'assets/img/edition/firm-signs/je/no6_lin20/firm_je_no6_lin20.png',
         short: '[JE] | No. 6 | 20 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 6 | 20 linig.',
     },
@@ -72,7 +72,7 @@ export const EDITION_FIRM_SIGNS_DATA = {
      * The path to the assets image of a firm sign: JE No. 9, 28 Lines.
      */
     FIRM_JE_NO_9_LIN_28: {
-        route: 'assets/img/edition/firm-signs/firm_je_no9_lin28.png',
+        route: 'assets/img/edition/firm-signs/je/no9_lin28/firm_je_no9_lin28.png',
         short: '[JE] | No. 9 | 28 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 9 | 28 linig.',
     },
@@ -81,7 +81,7 @@ export const EDITION_FIRM_SIGNS_DATA = {
      * The path to the assets image of a firm sign: JE No. 15, 16 Lines.
      */
     FIRM_JE_NO_15_LIN_16: {
-        route: 'assets/img/edition/firm-signs/firm_je_no15_lin16.png',
+        route: 'assets/img/edition/firm-signs/je/no15_lin16/firm_je_no15_lin16.png',
         short: '[JE] | No. 15 | 16 linig.',
         full: 'J. E. & Co. [diagonal ansteigend auf schildförmiger Papierrolle, rechts von einem Löwen gehalten, links und unten Blattgirlande] | Protokoll. Schutzmarke | No. 15 | 16 linig.',
     },
