@@ -62,22 +62,22 @@ export class SourceDescriptionComponent {
             B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_6_LIN_20],
         },
         M30: {
-            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18B],
+            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_B],
         },
         M31: {
-            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18B],
+            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_B],
         },
         M34: {
-            A: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18C],
-            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18D],
+            A: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_C],
+            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_D],
         },
         M35_42: {
-            A: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18C],
-            C: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18D],
+            A: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_C],
+            C: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_D],
         },
         M37: {
-            A: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18C],
-            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18D],
+            A: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_C],
+            B: [EDITION_FIRM_SIGNS_DATA.FIRM_JE_NO_5_LIN_18_D],
         },
     };
 
