@@ -64,11 +64,13 @@ describe('SourceEvaluationComponent (DONE)', () => {
         expectedComplexId = 'testComplex1';
         expectedNextComplexId = 'testComplex2';
         expectedFragment = 'source_A';
-        expectedSvgSheet = mockEditionData.mockSvgSheet_Sk1;
-        expectedNextSvgSheet = mockEditionData.mockSvgSheet_Sk2;
-        expectedModalSnippet = mockEditionData.mockModalSnippet;
-        expectedSourceEvaluationListData = { ...mockEditionData.mockSourceEvaluationListData };
-        expectedSourceEvaluationListEmptyData = { ...mockEditionData.mockSourceEvaluationListEmptyData };
+        expectedSvgSheet = JSON.parse(JSON.stringify(mockEditionData.mockSvgSheet_Sk1));
+        expectedNextSvgSheet = JSON.parse(JSON.stringify(mockEditionData.mockSvgSheet_Sk2));
+        expectedModalSnippet = JSON.parse(JSON.stringify(mockEditionData.mockModalSnippet));
+        expectedSourceEvaluationListData = JSON.parse(JSON.stringify(mockEditionData.mockSourceEvaluationListData));
+        expectedSourceEvaluationListEmptyData = JSON.parse(
+            JSON.stringify(mockEditionData.mockSourceEvaluationListEmptyData)
+        );
 
         mockDocument = TestBed.inject(DOCUMENT);
 
